vote/requirements.txt
Flask
Redis
gunicorn
psycopg2-binary
PyJWT
